Added status endpoint
