+++ requests/movements-add.rest
@@ -1,13 +1,13 @@
 POST http://localhost:3000/movements HTTP/1.1
 content-type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0OGIyYTQ4MDcyZDY5MWEyZTQ1NGZkNiIsInVzZXJuYW1lIjoicGlwZVVzZXI0IiwiaWF0IjoxNjg2ODQxOTM1LCJleHAiOjE2ODY4NDU1MzV9._T6Nujg8UKOGE-z82-ydDjB0ifjLdULqFHzhLebpMPU
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0OGIyZGY3ODE1MjAxYWY4ZGE4ZWIxMyIsInVzZXJuYW1lIjoicGlwZVVzZXIzIiwiaWF0IjoxNjg2ODQyODg4LCJleHAiOjE2ODY4NDY0ODh9.j1B-zEwKDjV8PPgiC49T8_SHFicIECoMJfylyLyoxY8
 
     {
       "name": "Coca",
       "description": "Coca con amigos",
       "type": "Expense",
       "category": "Oseo",
-      "date": "2015-03-25",
+      "date": "2016-03-25",
       "account": "Visa",
       "note": "No vale la pena",
       "amount": 100
