+++ requests/sign-in.rest
@@ -1,4 +1,4 @@
-POST http://localhost:3000/sign-in/ HTTP/1.1
+POST http://localhost:3000/user/sign-in HTTP/1.1
 content-type: application/json
 
 {
