+++ requests/sign-up.rest
@@ -1,4 +1,4 @@
-POST http://localhost:3000/sign-up HTTP/1.1
+POST http://localhost:3000/user/sign-up HTTP/1.1
 content-type: application/json
 
 {
