+++ requests/user-status.rest
@@ -1,0 +1,3 @@
+GET http://localhost:3000/user/status
+content-type: application/json
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0OGIyZGY3ODE1MjAxYWY4ZGE4ZWIxMyIsInVzZXJuYW1lIjoicGlwZVVzZXIzIiwiaWF0IjoxNjg2ODQyODg4LCJleHAiOjE2ODY4NDY0ODh9.j1B-zEwKDjV8PPgiC49T8_SHFicIECoMJfylyLyoxY8
