+++ src/app.js
@@ -1,7 +1,6 @@
 import express, { json } from 'express'
 import { connectDb } from './db/db.js'
-import { signInRouter } from './routes/signInRoutes.js'
-import { signUpRouter } from './routes/signUpRoutes.js'
+import { userRouter } from './routes/userRoutes.js'
 import { movementRouter } from './routes/movementRoutes.js'
 
 const app = express()
@@ -11,8 +10,7 @@ connectDb()
 
 app.use(json())
 
-app.use('/sign-up', signUpRouter)
-app.use('/sign-in', signInRouter)
+app.use('/user', userRouter)
 app.use('/movements', movementRouter)
 
 const server = app.listen(PORT, () => {
