+++ src/controllers/movementController.js
@@ -6,6 +6,14 @@ export const addMovement = async (req, res) => {
 
   const movement = new Movement({ name, description, type, category, note, user: user.id, account, date, amount })
   await movement.save()
+
+  if (type === 'Income') {
+    user.incomes = user.incomes + amount
+  } else {
+    // Expense
+    user.expenses = user.expenses + amount
+  }
+
   user.movements = user.movements.concat(movement._id)
   await user.save()
