+++ src/models/User.js
@@ -6,7 +6,14 @@ const userSchema = new mongoose.Schema({
   email: String,
   username: String,
   passwordHash: String,
-  
+  incomes: {
+    type: Number,
+    default: 0
+  },
+  expenses: {
+    type: Number,
+    default: 0
+  },
   movements: [{
     type: mongoose.Schema.Types.ObjectId,
     ref: 'Movement'
@@ -15,7 +22,6 @@ const userSchema = new mongoose.Schema({
 
 userSchema.set('toJSON', {
   transform: (document, returnObject) => {
-    returnObject.id = returnObject._id
     delete returnObject._id
     delete returnObject.__v
     delete returnObject.passwordHash
