--- src/routes/signInRoutes.js
@@ -1,8 +1,0 @@
-import express from 'express'
-import { signIn } from '../controllers/signInController.js'
-
-const router = express.Router()
-
-router.post('/', signIn)
-
-export { router as signInRouter }
