--- src/routes/signUpRoutes.js
@@ -1,8 +1,0 @@
-import express from 'express'
-import { singUp } from '../controllers/signUpController.js'
-
-const router = express.Router()
-
-router.post('/', singUp)
-
-export { router as signUpRouter }
